app/requirements.txt
fastapi
uvicorn
requests
numpy
opencv-python-headless
pillow
torch
torchvision
ultralytics
firebase-admin